+++ x-pack/plugins/security_solution/public/onboarding/components/onboarding_body/cards/integrations/__mocks__/integration_card_grid_tabs.tsx
@@ -7,4 +7,4 @@
 
 import React from 'react';
 
-export const PackageInstalledCallout = () => <div data-test-subj="packageInstalledCallout" />;
+export const IntegrationsCardGridTabs = () => <div data-test-subj="integrationsCardGridTabs" />;
