+++ x-pack/plugins/security_solution/public/onboarding/components/onboarding_body/cards/integrations/__mocks__/package_list_grid.tsx
@@ -4,7 +4,6 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-
 import React from 'react';
 
-export const AvailablePackages = () => <div data-test-subj="availablePackages" />;
+export const PackageListGrid = () => <div data-test-subj="packageListGrid" />;
