+++ x-pack/plugins/security_solution/public/onboarding/components/onboarding_body/cards/integrations/callouts/agent_required_callout.test.tsx
@@ -13,9 +13,9 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { AgentRequiredCallout } from './agent_required_callout';
-import { TestProviders } from '../../../../../common/mock/test_providers';
+import { TestProviders } from '../../../../../../common/mock/test_providers';
 
-jest.mock('../../../../../common/lib/kibana');
+jest.mock('../../../../../../common/lib/kibana');
 
 describe('AgentRequiredCallout', () => {
   beforeEach(() => {
